fixed rotation

+++ src/grid.cpp
@@ -37,7 +37,7 @@ void Grid::init(std::vector<std::vector<unsigned int>> tileData, unsigned int le
     unsigned int width = tileData[0].size();
     this->ROW = height;
     this->COL = width;
-    float unit_width = levelWidth / static_cast<float>(width);
+    float unit_width = levelWidth / width;
     float unit_height = levelHeight / height;
     this->unitWidth = unit_width;
     this->unitHeight = unit_height;
